+++ neuron/src/lib/Data/Graph/Labelled.hs
@@ -28,6 +28,7 @@ module Data.Graph.Labelled
     bfsForestBackwards,
     obviateRootUnlessForest,
     induceOnEdge,
+    induceOnEdgeReplacing,
     induce,
     emap,
     vmap,