+++ neuron/src/lib/Neuron/Zettelkasten/Connection.hs
@@ -14,11 +14,9 @@ import Text.Show (Show (show))
 
 -- | Represent the connection between zettels
 data Connection
-  = -- | A folgezettel points to a zettel that is conceptually a *part* of the
-    -- parent zettel.
-    Folgezettel
-  | -- | Any other ordinary connection (eg: "See also")
-    OrdinaryConnection
+  = Folgezettel
+  | FolgezettelInverse
+  | OrdinaryConnection
   deriving (Eq, Ord, Enum, Bounded, Generic)
 
 instance FromJSON Connection where
@@ -32,6 +30,8 @@ instance Default Connection where
 
 instance Semigroup Connection where
   -- A folgezettel link trumps all other kinds in that zettel.
+  FolgezettelInverse <> _ = FolgezettelInverse
+  _ <> FolgezettelInverse = FolgezettelInverse
   Folgezettel <> _ = Folgezettel
   _ <> Folgezettel = Folgezettel
   OrdinaryConnection <> OrdinaryConnection = OrdinaryConnection
@@ -39,11 +39,13 @@ instance Semigroup Connection where
 instance Show Connection where
   show = \case
     Folgezettel -> "folge"
+    FolgezettelInverse -> "folgeinv"
     OrdinaryConnection -> "cf"
 
 instance Read Connection where
   readsPrec _ s
     | s == show Folgezettel = [(Folgezettel, "")]
+    | s == show FolgezettelInverse = [(FolgezettelInverse, "")]
     | s == show OrdinaryConnection = [(OrdinaryConnection, "")]
     | otherwise = []