+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   - neuronignore: Exclude directories and markdown files via `.neuronignore`. (#499)
     - By default, recursively read Markdown files in notes directory. This behaviour can be disabled via the neuronignore plugin.
   - link queries & tags are now plugins
+- Plugin:links
+  - Labelled directional links (#514): `[[foo]]#` and `#[[foo]]`
 - Change default slug to use all lower-case (#527)
 - Reduce error verbosity (CLI & web) in the scenario of there being innumerous broken wiki-links.
 - CLI